+++ core/src/main/java/edu/ucr/cs/riple/core/Annotator.java
@@ -25,7 +25,6 @@
 package edu.ucr.cs.riple.core;
 
 import com.google.common.collect.ImmutableSet;
-import edu.ucr.cs.css.Serializer;
 import edu.ucr.cs.riple.core.explorers.BasicExplorer;
 import edu.ucr.cs.riple.core.explorers.ExhaustiveExplorer;
 import edu.ucr.cs.riple.core.explorers.Explorer;
@@ -40,6 +39,7 @@ import edu.ucr.cs.riple.core.metadata.trackers.CompoundTracker;
 import edu.ucr.cs.riple.core.metadata.trackers.RegionTracker;
 import edu.ucr.cs.riple.core.util.Utility;
 import edu.ucr.cs.riple.injector.Change;
+import edu.ucr.cs.scanner.Serializer;
 import java.util.HashMap;
 import java.util.Set;
 import java.util.stream.Collectors;
