+++ core/src/main/java/edu/ucr/cs/riple/core/metadata/trackers/CompoundTracker.java
@@ -24,9 +24,9 @@
 
 package edu.ucr.cs.riple.core.metadata.trackers;
 
-import edu.ucr.cs.css.Serializer;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
 import edu.ucr.cs.riple.core.metadata.method.MethodInheritanceTree;
+import edu.ucr.cs.scanner.Serializer;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
