+++ scanner/src/main/java/edu/ucr/cs/scanner/XMLUtil.java
@@ -4,7 +4,7 @@
  * MODIFIED TO REUSE IN THIS PROJECT.
  */
 
-package edu.ucr.cs.css;
+package edu.ucr.cs.scanner;
 
 import java.io.IOException;
 import java.nio.file.Files;
